Revert "Introduce ActionUriTrait to avoid duplicate code"
This reverts commit 64c28052eaea3ad1ea1a36543442474f8a2c05b7.
